modify: Nav 컴포넌트화
